refactor: Extract terminal data to separate file
